Improve instruction decoding and field extraction
Initialized Instruction struct to zero to avoid garbage values and added explicit extraction of rd, rn, rm, shift_amt, and imm fields in decode(). Enhanced get_opcode() logic for more accurate opcode detection and updated comments for clarity. Changed rd, rn, rm types in Instruction struct to uint8_t for correctness.

+++ src/decode.h
@@ -5,9 +5,9 @@
 typedef struct {
     Opcode opc;
     uint32_t hex;
-    int rd;
-    int rn;
-    int rm;
+    uint8_t rd;
+    uint8_t rn;
+    uint8_t rm;
     int imm;
     int shift_amt;
 } Instruction;
